# prisma/swagger.yaml
openapi: "3.0.0"
paths:
...
  /api/car/{id}:
    get:
      operationId: "CarController_getCar"
      summary: "Get car by ID"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Return car details."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDto"
        404:
          description: "Car not found."
      tags:
        - "car"
      security:
        - cookie: []
    delete:
      operationId: "CarController_deleteCar"
      summary: "Delete a car"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Car deleted successfully."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDto"
        404:
          description: "Car not found."
      tags:
        - "car"
      security:
        - cookie: []
...
  /api/vehicle-catalog/makes/{makeId}/models:
    get:
      operationId: "VehicleCatalogController_getModels"
      summary: "List models for a given make"
      parameters:
        - name: "makeId"
          required: true
          in: "path"
          schema:
            type: "number"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ModelResponseDto"
      tags:
        - "vehicle-catalog"
  /api/vehicle-catalog/models/{modelId}/years:
    get:
      operationId: "VehicleCatalogController_getYears"
      summary: "List years for a given model"
      parameters:
        - name: "modelId"
          required: true
          in: "path"
          schema:
            type: "number"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/YearResponseDto"
      tags:
        - "vehicle-catalog"
  /api/maintenance-record:
    post:
      operationId: "MaintenanceRecordController_createMaintenanceRecord"
      summary: "Create a new maintenance record"
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceRecordDto"
      responses:
        201:
          description: "The maintenance record has been successfully created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecordDto"
      tags:
        - "maintenance-record"
      security:
        - bearer: []
  /api/maintenance-record/{id}:
    get:
      operationId: "MaintenanceRecordController_getMaintenanceRecord"
      summary: "Get maintenance record by ID"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Return maintenance record details."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecordDto"
        404:
          description: "Maintenance record not found."
      tags:
        - "maintenance-record"
      security:
        - bearer: []
    put:
      operationId: "MaintenanceRecordController_updateMaintenanceRecord"
      summary: "Update a maintenance record"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMaintenanceRecordDto"
      responses:
        200:
          description: "The maintenance record has been successfully updated."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecordDto"
        404:
          description: "Maintenance record not found."
      tags:
        - "maintenance-record"
      security:
        - bearer: []
    delete:
      operationId: "MaintenanceRecordController_deleteMaintenanceRecord"
      summary: "Delete a maintenance record"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "The maintenance record has been successfully deleted."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecordDto"
        404:
          description: "Maintenance record not found."
      tags:
        - "maintenance-record"
      security:
        - bearer: []
  /api/maintenance-record/car/{carId}:
    get:
      operationId: "MaintenanceRecordController_getMaintenanceRecordsByCarId"
      summary: "Get all maintenance records for a specific car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Return list of maintenance records."
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/MaintenanceRecordDto"
      tags:
        - "maintenance-record"
      security:
        - bearer: []
  /api/document:
    post:
      operationId: "DocumentController_createDocument"
      summary: "Create a new document"
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDto"
      tags:
        - "document"
  /api/document/{id}:
    get:
      operationId: "DocumentController_getDocument"
      summary: "Get document by ID"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDto"
        404:
          description: "Document not found."
      tags:
        - "document"
    put:
      operationId: "DocumentController_updateDocument"
      summary: "Update a document"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDto"
        404:
          description: "Document not found."
      tags:
        - "document"
    delete:
      operationId: "DocumentController_deleteDocument"
      summary: "Delete a document"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDto"
      tags:
        - "document"
  /api/document/car/{carId}:
    get:
      operationId: "DocumentController_getDocumentsByCarId"
      summary: "Get all documents for a specific car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/DocumentDto"
      tags:
        - "document"
  /api/document/extract:
    post:
      operationId: "DocumentController_extractDocument"
      summary: "Extract fields from an uploaded document file (returns suggestions only — does not save)"
      description: "Upload a PDF document to have its fields extracted and returned as suggestions. No data is stored. The user must review and confirm the extracted values before saving. Currently supports Romanian RCA insurance PDFs."
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: "object"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionResultDto"
        400:
          description: "No file provided or file could not be read."
        415:
          description: "Unsupported file type."
      tags:
        - "document"
  /api/document/{id}/upload:
    post:
      operationId: "DocumentController_uploadFile"
      summary: "Upload a file for a document"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: "object"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDto"
      tags:
        - "document"
...
info:
  title: "History Auto Utility API"
  description: "API documentation for History Auto Utility"
  version: "1.0"
  contact: {}
...
servers:
  - url: "http://localhost:3000"
    description: "development"
components:
  securitySchemes:
    bearer:
      scheme: "bearer"
      bearerFormat: "JWT"
      type: "http"
  schemas:
...
    CarPhotoDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        url:
          type: "string"
          example: "/uploads/cars/photo.jpg"
        is_default:
          type: "boolean"
          example: false
      required:
        - "id"
        - "url"
        - "is_default"
    CarDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        user_id:
          type: "number"
          example: 123
        vin:
          type: "string"
          example: "1HGCM82633A123456"
        nickname:
          type: "string"
          example: "Blue Beast"
        make:
          type: "string"
          example: "BMW"
        model:
          type: "string"
          example: "530iA"
        variant:
          type: "string"
          example: "M Sport"
        year:
          type: "number"
          example: 1992
        license_plate:
          type: "string"
          example: "IS49WAU"
        fuel_type:
          type: "string"
          example: "PETROL"
          enum:
            - "PETROL"
            - "DIESEL"
            - "HYBRID"
            - "PLUGIN_HYBRID"
            - "ELECTRIC"
            - "LPG"
        transmission:
          type: "string"
          example: "MANUAL"
          enum:
            - "MANUAL"
            - "AUTOMATIC"
            - "SEMI_AUTOMATIC"
        engine:
          type: "string"
          example: "2.0L Turbo"
        color:
          type: "string"
          example: "#1A2B3C"
        current_mileage:
          type: "number"
          example: 184200
        ownership_start_date:
          format: "date-time"
          type: "string"
          example: "2024-05-12T00:00:00.000Z"
        rca_expiry_date:
          format: "date-time"
          type: "string"
          example: "2026-03-10T00:00:00.000Z"
        itp_expiry_date:
          format: "date-time"
          type: "string"
          example: "2026-04-20T00:00:00.000Z"
        rov_expiry_date:
          format: "date-time"
          type: "string"
          example: "2026-05-15T00:00:00.000Z"
        last_oil_service_date:
          format: "date-time"
          type: "string"
          example: "2026-03-12T00:00:00.000Z"
        last_oil_service_mileage:
          type: "number"
          example: 180000
        status:
          type: "string"
          example: "ACTIVE"
          enum:
            - "ACTIVE"
            - "SOLD"
        sold_at:
          format: "date-time"
          type: "string"
          example: "2025-06-01T00:00:00.000Z"
        photos:
          type: "array"
          items:
            $ref: "#/components/schemas/CarPhotoDto"
      required:
        - "id"
        - "user_id"
        - "vin"
        - "nickname"
        - "make"
        - "model"
        - "variant"
        - "year"
        - "license_plate"
        - "fuel_type"
        - "transmission"
        - "engine"
        - "color"
        - "ownership_start_date"
        - "rca_expiry_date"
        - "itp_expiry_date"
        - "rov_expiry_date"
        - "last_oil_service_date"
        - "last_oil_service_mileage"
        - "status"
        - "photos"
...
    ModelResponseDto:
      type: "object"
      properties:
        id:
          type: "number"
        makeId:
          type: "number"
        name:
          type: "string"
        normalizedName:
          type: "string"
      required:
        - "id"
        - "makeId"
        - "name"
        - "normalizedName"
    YearResponseDto:
      type: "object"
      properties:
        id:
          type: "number"
        modelId:
          type: "number"
        year:
          type: "number"
      required:
        - "id"
        - "modelId"
        - "year"
    CreateMaintenanceRecordDto:
      type: "object"
      properties:
        car_id:
          type: "number"
          example: 123
        service_date:
          type: "string"
          example: "2023-01-01"
        mileage:
          type: "number"
          example: 51000
        description:
          type: "string"
          example: "Oil 10W40, air filter MANN"
        service_type:
          type: "string"
          example: "MAINTENANCE"
          enum:
            - "REPAIR"
            - "MAINTENANCE"
            - "IMPROVEMENT"
        service_category:
          type: "string"
          example: "OIL_CHANGE"
          enum:
            - "OIL_CHANGE"
            - "BRAKE_SERVICE"
            - "TRANSMISSION_SERVICE"
            - "TIRE_SERVICE"
            - "FLUID_SERVICE"
            - "ENGINE_SERVICE"
            - "INSPECTION"
            - "BATTERY_SERVICE"
            - "FILTER_SERVICE"
            - "LIGHT_SERVICE"
            - "OTHER"
        cost:
          type: "number"
          example: 250
        expiry_date:
          type: "string"
          example: "2024-01-01"
      required:
        - "car_id"
        - "service_date"
        - "mileage"
        - "description"
        - "service_type"
        - "service_category"
        - "cost"
    MaintenanceRecordDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: "123"
        car_id:
          type: "number"
          example: "123"
        service_date:
          format: "date-time"
          type: "string"
          example: "2023-01-01T00:00:00.000Z"
        mileage:
          type: "number"
          example: "380000"
        description:
          type: "string"
          example: "Oil 10W40, air filter MANN"
        service_type:
          type: "string"
          example: "MAINTENANCE"
          enum:
            - "REPAIR"
            - "MAINTENANCE"
            - "IMPROVEMENT"
        service_category:
          type: "string"
          example: "OIL_CHANGE"
          enum:
            - "OIL_CHANGE"
            - "BRAKE_SERVICE"
            - "TRANSMISSION_SERVICE"
            - "TIRE_SERVICE"
            - "FLUID_SERVICE"
            - "ENGINE_SERVICE"
            - "INSPECTION"
            - "BATTERY_SERVICE"
            - "FILTER_SERVICE"
            - "LIGHT_SERVICE"
            - "OTHER"
        cost:
          type: "number"
          example: "Insurance"
        expiry_date:
          format: "date-time"
          type: "string"
          example: "Insurance"
      required:
        - "id"
        - "car_id"
        - "service_date"
        - "mileage"
        - "description"
        - "service_type"
        - "service_category"
        - "cost"
        - "expiry_date"
    UpdateMaintenanceRecordDto:
      type: "object"
      properties: {}
    CreateDocumentDto:
      type: "object"
      properties:
        document_type:
          type: "string"
          example: "RCA"
        car_id:
          type: "number"
          example: 1
        issue_date:
          type: "string"
          example: "2023-01-01"
        expiry_date:
          type: "string"
          example: "2024-01-01"
        provider:
          type: "string"
          example: "Allianz-Tiriac Asigurări"
        policy_number:
          type: "string"
          example: "POL-123456789"
        status:
          type: "string"
          example: "Active"
        policyholder:
          type: "string"
          example: "John Doe"
        cnp_id:
          type: "string"
          example: "1900101012345"
      required:
        - "document_type"
        - "car_id"
    DocumentDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        document_type:
          type: "string"
          example: "RCA"
        car_id:
          type: "number"
          example: 1
        issue_date:
          format: "date-time"
          type: "string"
          nullable: true
        expiry_date:
          format: "date-time"
          type: "string"
          nullable: true
        provider:
          type: "string"
          nullable: true
        policy_number:
          type: "string"
          nullable: true
        status:
          type: "string"
          nullable: true
        policyholder:
          type: "string"
          nullable: true
        cnp_id:
          type: "string"
          nullable: true
        file_url:
          type: "string"
          nullable: true
        file_name:
          type: "string"
          nullable: true
        file_size:
          type: "number"
          nullable: true
      required:
        - "id"
        - "document_type"
        - "car_id"
    UpdateDocumentDto:
      type: "object"
      properties: {}
    ExtractedFieldsDto:
      type: "object"
      properties:
        policy_series:
          type: "string"
          description: "Insurance policy series"
        policy_number:
          type: "string"
          description: "Insurance policy number"
        insurer_name:
          type: "string"
          description: "Insurer / insurance company name"
        broker_name:
          type: "string"
          description: "Broker or agent name"
        policyholder_name:
          type: "string"
          description: "Policyholder full name"
        owner_name:
          type: "string"
          description: "Owner name, if different from policyholder"
        owner_cnp:
          type: "string"
          description: "Owner Romanian personal ID (CNP)"
        plate_number:
          type: "string"
          description: "Vehicle license plate number"
        vin:
          type: "string"
          description: "Vehicle Identification Number (VIN / chassis number)"
        vehicle_make:
          type: "string"
          description: "Vehicle make / brand"
        vehicle_model:
          type: "string"
          description: "Vehicle model"
        vehicle_category:
          type: "string"
          description: "Vehicle category (e.g. M1, N1)"
        engine_capacity:
          type: "string"
          description: "Engine displacement in cc"
        power:
          type: "string"
          description: "Engine power in kW"
        seats:
          type: "string"
          description: "Number of seats"
        max_weight:
          type: "string"
          description: "Maximum authorized weight"
        valid_from:
          type: "string"
          description: "Policy valid from date (ISO 8601)"
        valid_until:
          type: "string"
          description: "Policy valid until date (ISO 8601)"
        issue_date:
          type: "string"
          description: "Document issue date (ISO 8601)"
        premium:
          type: "string"
          description: "Insurance premium amount as a numeric string"
        currency:
          type: "string"
          description: "Currency code (RON, EUR, USD)"
        bonus_malus_class:
          type: "string"
          description: "Bonus-malus class (e.g. B0, B5, M1)"
        direct_settlement:
          type: "boolean"
          description: "Whether direct settlement (decontare directă) is included"
        direct_settlement_price:
          type: "string"
          description: "Direct settlement limit/price if available"
        payment_installments:
          type: "string"
          description: "Payment installments information"
        damage_limits:
          type: "string"
          description: "Damage limits (despăgubire maximă)"
    ExtractionResultDto:
      type: "object"
      properties:
        detected:
          type: "boolean"
          description: "Whether the document was recognised as a supported type"
        document_type:
          type: "string"
          description: "Detected document type, e.g. RCA"
          example: "RCA"
        confidence:
          type: "string"
          description: "Extraction confidence level"
          enum:
            - "high"
            - "medium"
            - "low"
            - "none"
          example: "high"
        fields:
          description: "Extracted field suggestions for frontend pre-fill"
          allOf:
            - $ref: "#/components/schemas/ExtractedFieldsDto"
        warnings:
          description: "Fields or issues that require manual review"
          type: "array"
          items:
            type: "string"
        raw_text:
          type: "string"
          description: "First 3000 characters of raw extracted text, useful for debugging"
      required:
        - "detected"
        - "confidence"
        - "fields"
        - "warnings"
...